massive store and types update
