+++ src/containers/BrokerSelectContainer.ts
@@ -1,7 +1,7 @@
 import { connect, ConnectedProps } from "react-redux";
 import { BrokerSelect } from "../components/BrokerSelect";
-import { IRootState } from "../types/redux";
-import { updateBrokerType } from "../redux/actionCreators/uiActions";
+import { IRootState } from "../redux/store";
+import { updateBroker } from "../redux/slices/ui";
 
 const mapStateToBrokerSelectProps = ({ ui }: IRootState) => {
   const { broker } = ui;
@@ -11,7 +11,7 @@ const mapStateToBrokerSelectProps = ({ ui }: IRootState) => {
   };
 };
 
-const connector = connect(mapStateToBrokerSelectProps, { updateBrokerType });
+const connector = connect(mapStateToBrokerSelectProps, { updateBroker });
 
 export type IBrokerSelectContainer = ConnectedProps<typeof connector>;
