+++ src/containers/DataSubsectionContainer.tsx
@@ -1,19 +1,19 @@
 import { connect, ConnectedProps } from "react-redux";
 import { DataSubsection } from "../components/DataSubsection";
-import { BROKERS } from "../redux/reducers/ui";
-import { IRootState } from "../types/redux";
+import { BROKERS } from "../redux/slices/ui";
+import { IRootState } from "../redux/store";
 
 export interface ISubsectionOwnProps {
-  datasetProp: string;
-  totalProp: string;
+  datasetProp: keyof IRootState["data"]["parsedData"];
+  totalProp: keyof IRootState["data"]["parsedData"];
 }
 
 const mapStateToSubsectionProps = (
   { data, ui }: IRootState,
   { totalProp, datasetProp }: ISubsectionOwnProps
 ) => {
-  const dataset = data[datasetProp] as string;
-  const total = data[totalProp] as number;
+  const dataset = data.parsedData[datasetProp] as string;
+  const total = data.parsedData[totalProp] as number;
 
   const { broker } = ui;
   const isHidden = broker !== BROKERS.EXANTE && datasetProp === "interests";
