--- src/redux/actionCreators/dataActions.ts
@@ -1,7 +1,0 @@
-import { IDataStore, IUpdateDataAction } from "../../types/redux";
-import { UPDATE_DATA_STORE } from "../actionTypes";
-
-export const updateDataStore = (data: IDataStore): IUpdateDataAction => ({
-  type: UPDATE_DATA_STORE,
-  payload: { ...data },
-});
