--- src/redux/actionCreators/uiActions.ts
@@ -1,8 +1,0 @@
-import { IUpdateUiAction } from "../../types/redux";
-import { IBrokerTypes } from "../../types/types";
-import { UPDATE_BROKER_TYPE } from "../actionTypes";
-
-export const updateBrokerType = (type: IBrokerTypes): IUpdateUiAction => ({
-  type: UPDATE_BROKER_TYPE,
-  payload: type,
-});
