--- src/redux/actionTypes.ts
@@ -1,2 +1,0 @@
-export const UPDATE_DATA_STORE = "UPDATE_DATA_STORE";
-export const UPDATE_BROKER_TYPE = "UPDATE_BROKER_TYPE";
